% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_prefix}
\alias{add_prefix}
\title{Add a prefix to each word}
\usage{
add_prefix(texts, prefix)
}
\arguments{
\item{texts}{a \link{character} containing the original text}

\item{prefix}{unit \link{character} containing the prefix to add (length == 1) or \link{character} with same length than texts}
}
\value{
\link{character} with prefixed words.
}
\description{
Add a custom prefix to each word of a a line to create different spaces.
Code in C++ (efficient).
}
\examples{
add_prefix(c("this is a test", "this is another    test"), "#")
}
