% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{build_vectors}
\alias{build_vectors}
\title{Build fasttext vectors}
\usage{
build_vectors(documents, description, model_path, modeltype = c("skipgram",
  "cbow"), lr = 0.05, dim = 100, ws = 5, epoch = 5, minCount = 5,
  minCountLabel = 0, neg = 5, wordNgrams = 1, loss = c("ns", "hs",
  "softmax"), bucket = 2e+06, minn = 3, maxn = 6, thread = 12,
  lrUpdateRate = 100, t = 1e-04, label = "__label__", verbose = 2,
  pretrainedVectors = NULL)
}
\arguments{
\item{documents}{character vector of documents used for training}

\item{lr}{learning rate}

\item{dim}{size of word vectors}

\item{ws}{size of the context window}

\item{epoch}{number of epochs}

\item{minCount}{minimal number of word occurences}

\item{neg}{number of negatives sampled}

\item{wordNgrams}{max length of word ngram}

\item{loss}{loss function {ns, hs, softmax}}

\item{bucket}{number of buckets}

\item{minn}{min length of char ngram}

\item{maxn}{max length of char ngram}

\item{thread}{number of threads}

\item{lrUpdateRate}{change the rate of updates for the learning rate}

\item{t}{sampling threshold}

\item{targets}{vector of targets/catagory of each document. Must have same length as \code{documents} and be coercable to character#' @param modeltype character of either 'skipgram' or 'cbow'}
}
\value{
path to model file, as character
}
\description{
Trains a fasttext vector/unsupervised model following method described in
\href{https://arxiv.org/abs/1607.04606}{Enriching Word Vectors with Subword Information}
using the \href{https://fasttext.cc/}{fasttext} implementation.
}
\details{
See \href{https://fasttext.cc/docs/en/unsupervised-tutorial.html} for more information on
training unsupervised models using fasttext
}
