% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_hamming_loss}
\alias{get_hamming_loss}
\title{Hamming loss}
\usage{
get_hamming_loss(labels, predictions)
}
\arguments{
\item{labels}{list of labels}

\item{predictions}{list returned by the predict command (including both the probability and the categories)}
}
\description{
Compute the hamming loss. When there is only one category, this measure the accuracy.
}
