% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_labels}
\alias{get_labels}
\title{Get list of labels (supervised model)}
\usage{
get_labels(model)
}
\arguments{
\item{model}{trained Fasttext model}
}
\description{
Get a [character] containing each label seen during training.
}
