% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_nn}
\alias{get_nn}
\title{Get nearest neighbour vectors}
\usage{
get_nn(model, word, k)
}
\arguments{
\item{model}{trained Fasttext model. Null if train a new model.}

\item{word}{reference word}

\item{k}{\link{integer} defining the number of resutls to return}
}
\value{
\link{numeric} with distances with \link{names} as words
}
\description{
Find the \code{k} words with the smallest distance.
First execution can be slow because of some precomputations.
Search is done linearly, if your model is big you may want to use an approximate neighbour algorithm from other R packages (like RcppAnnoy).
}
