% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_parameters}
\alias{get_parameters}
\title{Export hyper parameters}
\usage{
get_parameters(model)
}
\arguments{
\item{model}{trained fasttext model}
}
\value{
\link{list} contaning each parameter
}
\description{
Retrieve hyper paramters used to train the model
}
\examples{

library(FastRText)
model_test_path <- system.file("extdata", "model_classification_test.bin", package = "FastRText")
model <- load_model(model_test_path)
print(head(get_parameters(model), 5))

}
