% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_word_vector}
\alias{get_word_vector}
\title{Get word embedding}
\usage{
get_word_vector(model, words)
}
\arguments{
\item{model}{trained Fasttext model}

\item{words}{[character] of words}
}
\description{
Return the vector representation of a word (unsupervised training)
or a label (supervised training).
}
