% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{predict.Rcpp_FastRText}
\alias{predict.Rcpp_FastRText}
\title{Get predictions (for supervised model)}
\usage{
\method{predict}{Rcpp_FastRText}(object, sentences, k = 1, ...)
}
\arguments{
\item{object}{trained Fasttext model}

\item{sentences}{[character] containing the sentences}

\item{k}{will return the k most probable labels (default = 1)}

\item{...}{not used}
}
\description{
Return probabilities for the sentences to be associated with K labels.
}
\examples{

library(FastRText)
data("test_sentences")
model_test_path <- system.file("extdata", "model_classification_test.bin", package = "FastRText")
model <- load_model(model_test_path)
sentence <- test_sentences[1, "text"]
print(predict(model, sentence))

}
