% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{predict}
\alias{predict}
\title{Get predictions (for supervised model)}
\usage{
predict(model, sentences, k = 1)
}
\arguments{
\item{model}{trained Fasttext model}

\item{sentences}{[character] containing the sentences}

\item{k}{will return the k most probable labels (default = 1)}
}
\description{
Return probabilities for the sentences to be associated with K labels.
}
