% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{print_help}
\alias{print_help}
\title{Print help}
\usage{
print_help()
}
\description{
Print command information, mainly to use with \code{\link[=execute]{execute()}} \code{function}.
}
\examples{

print_help()

}
