% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Rcpp_FastRText-class}
\alias{Rcpp_FastRText-class}
\title{Rcpp_FastRText class.}
\description{
Some details about this class and my plans for it in the body.
}
\section{Slots}{

\describe{
\item{\code{load}}{Load a model}

\item{\code{predict}}{Make a prediction}

\item{\code{execute}}{Execute commands}

\item{\code{get_vectors}}{Get vectors related to provided words}

\item{\code{get_parameters}}{Get parameters used to train the model}

\item{\code{get_words}}{List all words learned}

\item{\code{get_labels}}{List all labels learned}
}}

