% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{execute}
\alias{execute}
\title{Execute command on Fasttext model (including training)}
\usage{
execute(model = NULL, path = NULL, commands)
}
\arguments{
\item{model}{trained Fasttext model. Null if train a new model.}

\item{path}{path to an existing model. Null id command to be applied to an existing model.}

\item{commands}{[character] of commands}
}
\description{
Use the same commands than the one to use for the command line.
}
