% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{fastrtext}
\alias{fastrtext}
\alias{fastrtext-package}
\alias{fastrtext-package}
\title{fastrtext: FastText R Wrapper for Text Classification and Word Representation}
\description{
This package is an R wrapper of fastText by Facebook.
FastText is an open-source, free, lightweight library that allows
users to learn text representations and text classifiers.
It transforms text into continuous vectors that can later
be used on many language related task.
It works on standard, generic hardware (no GPU required).
It also includes model size reduction feature.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pommedeterresautee/FastRText}
  \item Report bugs at \url{https://github.com/pommedeterresautee/FastRText/issues}
}

}
\author{
\strong{Maintainer}: Michaël Benesty \email{michael@benesty.fr}

}
