% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{FastRText}
\alias{FastRText}
\alias{FastRText-package}
\alias{FastRText-package}
\title{FastRText: FastRText}
\description{
R wrapper of fastText by Facebook for both text classification and word reprensentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pommedeterresautee/FastRText}
  \item Report bugs at \url{https://github.com/pommedeterresautee/FastRText/issues}
}

}
\author{
\strong{Maintainer}: Michaël Benesty \email{michael@benesty.fr}

}
