% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{fastrtext}
\alias{fastrtext}
\alias{fastrtext-package}
\alias{fastrtext-package}
\title{fastrtext: Fastrtext for text classification and word representation}
\description{
fastrtext is an R wrapper of fastText by Facebook for both performing text classification and generating word dense vectorial representations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pommedeterresautee/FastRText}
  \item Report bugs at \url{https://github.com/pommedeterresautee/FastRText/issues}
}

}
\author{
\strong{Maintainer}: Michaël Benesty \email{michael@benesty.fr}

}
