% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{fastrtext}
\alias{fastrtext}
\alias{fastrtext-package}
\alias{fastrtext-package}
\title{fastrtext: 'fastText' Wrapper for Text Classification and Word Representation}
\description{
Learning text representations and a text classifier may rely
on the same family of technics. 'fastText' is an open-source, free,
lightweight library that allows users to perform both tasks.
It transforms text into continuous vectors that can later
be used on many language related task.
It works on standard, generic hardware (no 'GPU' required).
It also includes model size reduction feature.
'fastText' original source code is available
at <https://github.com/facebookresearch/fastText>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pommedeterresautee/fastrtext}
  \item \url{https://pommedeterresautee.github.io/fastrtext/}
  \item Report bugs at \url{https://github.com/pommedeterresautee/fastrtext/issues}
}

}
\author{
\strong{Maintainer}: Michaël Benesty \email{michael@benesty.fr}

Other contributors:
\itemize{
  \item Facebook, Inc \email{bojanowski@fb.com} [copyright holder]
}

}
