% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_dictionary}
\alias{get_dictionary}
\title{Get list of known words}
\usage{
get_dictionary(model)
}
\arguments{
\item{model}{trained Fasttext model}
}
\description{
Get a [character] containing each word seen during training.
}
\examples{

library(FastRText)
model_test_path <- system.file("extdata", "model_classification_test.bin", package = "FastRText")
model <- load_model(model_test_path)
print(head(get_dictionary(model), 5))

}
