% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_parameters}
\alias{get_parameters}
\title{Export hyper parameters}
\usage{
get_parameters(model)
}
\arguments{
\item{model}{trained fasttext model}
}
\description{
Get hyper paramters used to train the model
}
