% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_word_distance}
\alias{get_word_distance}
\title{Distance between two words}
\usage{
get_word_distance(model, w1, w2)
}
\arguments{
\item{model}{trained Fasttext model. Null if train a new model.}

\item{w1}{first word to compare}

\item{w2}{second word to compare}
}
\description{
Distance is equal to `1 - cosine`
}
