% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{load_model}
\alias{load_model}
\title{Load an existing Fasttext trained model}
\usage{
load_model(path)
}
\arguments{
\item{path}{path to the existing model}
}
\description{
Load and return a pointer to an existing model which will be used in other functions of this package.
}
\examples{

library(fastrtext)
model_test_path <- system.file("extdata", "model_classification_test.bin", package = "fastrtext")
model <- load_model(model_test_path)

}
