% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{predict.Rcpp_FastRText}
\alias{predict.Rcpp_FastRText}
\title{Get predictions (for supervised model)}
\usage{
\method{predict}{Rcpp_FastRText}(object, sentences, k = 1, ...)
}
\arguments{
\item{object}{trained Fasttext model}

\item{sentences}{[character] containing the sentences}

\item{k}{will return the k most probable labels (default = 1)}

\item{...}{not used}
}
\description{
Return probabilities for the sentences to be associated with K labels.
}
