% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Rcpp_FastRText-class}
\alias{Rcpp_FastRText-class}
\title{Rcpp_FastRText class}
\description{
Models are objects with several methods which can be called that way: model$method()
}
\section{Slots}{

\describe{
\item{\code{load}}{Load a model}

\item{\code{predict}}{Make a prediction}

\item{\code{execute}}{Execute commands}

\item{\code{get_vectors}}{Get vectors related to provided words}

\item{\code{get_parameters}}{Get parameters used to train the model}

\item{\code{get_dictionary}}{List all words learned}

\item{\code{get_labels}}{List all labels learned}
}}

